+++ entrypoint.sh
@@ -46,36 +46,6 @@ fi
 
 echo -e "\e[34m[Info] Using config file at: '$CONFIG_PATH'.\e[0m"
 
-# Check if GITHUB_TOKEN is set
-if [ -n "$GITHUB_TOKEN" ]; then
-    echo "$GITHUB_TOKEN" > "$HOME/.github-token"
-    chmod 600 "$HOME/.github-token"
-
-    # Configure Git with the provided GITHUB_TOKEN
-    echo -e "\e[34m[Info] Configuring GitHub credentials with hostname '$GITHUB_HOSTNAME'.\e[0m"
-    echo "machine ${GITHUB_HOSTNAME}
-    login oauth
-    password ${GITHUB_TOKEN}" >> "$HOME/.netrc"
-    chmod 600 "$HOME/.netrc"
-else
-    echo -e "\e[34m[Info] Private GitHub repositories will not be indexed since GITHUB_TOKEN was not set.\e[0m"
-fi
-
-# Check if GITLAB_TOKEN is set
-if [ -n "$GITLAB_TOKEN" ]; then
-    echo "$GITLAB_TOKEN" > "$HOME/.gitlab-token"
-    chmod 600 "$HOME/.gitlab-token"
-   
-    # Configure Git with the provided GITLAB_TOKEN
-    echo -e "\e[34m[Info] Configuring GitLab credentials with hostname '$GITLAB_HOSTNAME'.\e[0m"
-    echo "machine ${GITLAB_HOSTNAME}
-    login oauth
-    password ${GITLAB_TOKEN}" >> "$HOME/.netrc"
-    chmod 600 "$HOME/.netrc"
-else
-    echo -e "\e[34m[Info] GitLab repositories will not be indexed since GITLAB_TOKEN was not set.\e[0m"
-fi
-
 # Update nextjs public env variables w/o requiring a rebuild.
 # @see: https://phase.dev/blog/nextjs-public-runtime-variables/
 
@@ -89,7 +59,7 @@ if [ -z "$NEXT_PUBLIC_SOURCEBOT_VERSION" ] && [ ! -z "$SOURCEBOT_VERSION" ]; the
     export NEXT_PUBLIC_SOURCEBOT_VERSION="$SOURCEBOT_VERSION"
 fi
 
-find /app/public /app/.next -type f -name "*.js" |
+find /app/packages/web/public /app/packages/web/.next -type f -name "*.js" |
 while read file; do
     sed -i "s|BAKED_NEXT_PUBLIC_SOURCEBOT_TELEMETRY_DISABLED|${NEXT_PUBLIC_SOURCEBOT_TELEMETRY_DISABLED}|g" "$file"
     sed -i "s|BAKED_NEXT_PUBLIC_SOURCEBOT_VERSION|${NEXT_PUBLIC_SOURCEBOT_VERSION}|g" "$file"